requirements.txt
flask==3.0.2
flask-login==0.6.3
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
werkzeug==3.0.1
sqlalchemy==2.0.27
gunicorn==21.2.0